+++ src/input_context/context_instance.rs
@@ -199,9 +199,9 @@ impl ActionBind {
     ///
     /// The order of arguments follows the common "WASD" mapping.
     pub fn with_xy_axis<I: Into<Input>>(&mut self, up: I, left: I, down: I, right: I) -> &mut Self {
-        self.to(InputBind::new(up).with_modifier(SwizzleAxis::YXZ))
-            .to(InputBind::new(left).with_modifier(Negate::default()))
-            .to(InputBind::new(down)
+        self.to(up.with_modifier(SwizzleAxis::YXZ))
+            .to(left.with_modifier(Negate::default()))
+            .to(down
                 .with_modifier(Negate::default())
                 .with_modifier(SwizzleAxis::YXZ))
             .to(right)
@@ -210,7 +210,8 @@ impl ActionBind {
     /// Maps the given stick as 2-dimentional input.
     pub fn with_stick(&mut self, stick: GamepadStick) -> &mut Self {
         self.to(stick.x())
-            .to(InputBind::new(stick.y()).with_modifier(SwizzleAxis::YXZ))
+            .to(stick.y())
+            .with_modifier(SwizzleAxis::YXZ)
     }
 
     /// Adds action-level modifier.
@@ -384,25 +385,36 @@ impl InputBind {
             ignored: true,
         }
     }
-
-    /// Adds modifier.
-    #[must_use]
-    pub fn with_modifier(mut self, modifier: impl InputModifier) -> Self {
-        self.modifiers.push(Box::new(modifier));
-        self
-    }
-
-    /// Adds condition.
-    #[must_use]
-    pub fn with_condition(mut self, condition: impl InputCondition) -> Self {
-        self.conditions.push(Box::new(condition));
-        self
-    }
 }
 
 impl<I: Into<Input>> From<I> for InputBind {
     fn from(input: I) -> Self {
         Self::new(input)
+    }
+}
+
+/// A trait to ergonomically add a modifier or condition to any type that can be converted into a binding.
+pub trait InputBindModCond {
+    /// Adds modifier.
+    #[must_use]
+    fn with_modifier(self, modifier: impl InputModifier) -> InputBind;
+
+    /// Adds condition.
+    #[must_use]
+    fn with_condition(self, condition: impl InputCondition) -> InputBind;
+}
+
+impl<T: Into<InputBind>> InputBindModCond for T {
+    fn with_modifier(self, modifier: impl InputModifier) -> InputBind {
+        let mut binding = self.into();
+        binding.modifiers.push(Box::new(modifier));
+        binding
+    }
+
+    fn with_condition(self, condition: impl InputCondition) -> InputBind {
+        let mut binding = self.into();
+        binding.conditions.push(Box::new(condition));
+        binding
     }
 }